fix flow-type issues and coding style
