+++ src/main/d3utils.js
@@ -20,6 +20,20 @@ function getTrackScale(range: GenomeRange, width: number) {
   return scale;
 }
 
+/**
+ * Formats the size of a view and infers what prefix/unit to show.
+ * This formatting follows IGV's conventions regarding range display:
+ *  "1 bp", "101 bp", "1,001 bp", "1,001 kbp", ...
+ */
+function formatRange(viewSize: number): any {
+  var tmpViewSize = viewSize / 1000,
+      fprefix = d3.formatPrefix(Math.max(1, tmpViewSize)),
+      unit = fprefix.symbol + "bp",  // bp, kbp, Mbp, Gbp
+      prefix = d3.format(',f.0')(fprefix.scale(viewSize));
+  return {prefix, unit};
+}
+
 module.exports = {
+  formatRange,
   getTrackScale
 };
